server/requirements.txt
django==4.1.5
pandas==2.1.3
scikit-learn==1.3.2
requests==2.31.0
django-extensions==3.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1